Default parameter ranges for TUWmodel

Input parameter ranges as given in the TUWmodel documentation for param. Additionally includes tcalt and pcalt as used in hbv_pso, set to 0 (no pcalt/tcalt applied).

tuwmodel_params_default

Format

A matrix with two columns, min and max, and the following parameters as rows:

SCF snow correction factor (0.9-1.5);

DDF degree day factor (0.0-5.0 mm/degC/timestep);

Tr threshold temperature above which precipitation is rain (1.0-3.0 degC);

Ts threshold temperature below which precipitation is snow (-3.0-1.0 degC);

Tm threshold temperature above which melt starts (-2.0-2.0 degC);

LPrat parameter related to the limit for potential evaporation (0.0-1.0);

FC field capacity, i.e., max soil moisture storage (0-600 mm);

BETA the non linear parameter for runoff production (0.0-20.0);

k0 storage coefficient for very fast response (0.0-2.0 timestep);

k1 storage coefficient for fast response (2.0-30.0 timestep);

k2 storage coefficient for slow response (30.0-250.0 timestep);

lsuz threshold storage state, i.e., the very fast response start if exceeded (1.0-100.0 mm);

cperc constant percolation rate (0.0-8.0 mm/timestep);

bmax maximum base at low flows (0.0-30.0 timestep);

croute free scaling parameter (0.0-50.0 timestep2/mm);

tcalt Lapse rate to adjust the temperature data by elevation zone (ºC/100m, decreasing with elevation)

pcalt Lapse rate to adjust the precipitation data by elevation zone (